pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dtce-ai-bot"
version = "1.0.0"
description = "Internal AI assistant Teams bot for DTCE engineering files"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "DTCE Engineering Team", email = "engineering@dtce.com"}
]
maintainers = [
    {name = "DTCE Engineering Team", email = "engineering@dtce.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business :: Office Suites",
]
keywords = ["teams", "bot", "ai", "assistant", "microsoft", "azure", "openai"]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "botbuilder-core>=4.15.0",
    "botbuilder-schema>=4.15.0",
    "botbuilder-integration-aiohttp>=4.15.0",
    "aiohttp>=3.8.5",
    "aiohttp-cors>=0.7.0",
    "msal>=1.25.0",
    "requests>=2.31.0",
    "azure-storage-blob>=12.19.0",
    "azure-search-documents>=11.4.0",
    "azure-cognitiveservices-language-textanalytics>=5.3.0",
    "openai>=1.3.7",
    "python-docx>=1.1.0",
    "PyPDF2>=3.0.1",
    "openpyxl>=3.1.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=23.2.0",
    "rich>=13.7.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0"
]

[project.scripts]
dtce-ai-bot = "app:main"

[project.urls]
"Homepage" = "https://github.com/dtce/ai-bot"
"Bug Reports" = "https://github.com/dtce/ai-bot/issues"
"Source" = "https://github.com/dtce/ai-bot"

[tool.setuptools.packages.find]
where = ["."]
include = ["dtce_ai_bot*"]
exclude = ["tests*", "scripts*", "docs*"]

[tool.setuptools.package-data]
dtce_ai_bot = ["teams-app/*.json", "teams-app/*.png"]

# Development tools configuration
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["dtce_ai_bot"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["dtce_ai_bot"]
omit = [
    "*/tests/*",
    "*/venv/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
